re-arrage files to group e.g. cli related stuff together update tests
