+++ chanjo/cli/calculate.py
@@ -1,13 +1,22 @@
 # -*- coding: utf-8 -*-
+import json
 import logging
 
 import click
 
 from chanjo.store.api import ChanjoDB
 from chanjo.store.constants import STAT_COLUMNS
-from .utils import dump_json
 
-log = logging.getLogger(__name__)
+LOG = logging.getLogger(__name__)
+
+
+def dump_json(data, pretty=False):
+    """Print JSON to console."""
+    if pretty:
+        json_args = dict(indent=4, sort_keys=True)
+    else:
+        json_args = {}
+    return json.dumps(data, **json_args)
 
 
 @click.group()
