+++ setup.py
@@ -95,12 +95,12 @@ setup(
         ],
         'chanjo.subcommands.4': [
             'init = chanjo.init.cli:init',
-            'sex = chanjo.sex.cli:sex',
+            'sex = chanjo.cli:sex',
             'sambamba = chanjo.sambamba.cli:sambamba',
             'db = chanjo.store.cli:db',
-            'load = chanjo.load.cli:load',
-            'link = chanjo.load.cli:link',
-            'calculate = chanjo.calculate.cli:calculate',
+            'load = chanjo.cli:load',
+            'link = chanjo.cli:link',
+            'calculate = chanjo.cli:calculate',
         ]
     },
